:atom: Initial alpha release of OkadaPy
This release provides a minimally complete implementation of the Okada
equations, with tooling provided to parse Coulomb3.3-formatted input
files and compute displacement, strain, and stress fields for the input
sources. There are, notably, a couple of differences to the
implementation when compared to the version of Coulomb available from
the USGS website.

1. The strain is now correctly computed (there was an issue with
   Coulomb)
2. The displacements for a point source of deflation/inflation differ by
   a factor of x1000, as I believe these were not being computed
correctly in Coulomb. I may be wrong.

Note: This package is by no means a complete replacement for Coulomb.
Many of the functions are not implemented, such as computing the Coulomb
stress on receiver faults etc. These might be added one day, but I was
primarily interested in just being able to model the displacement,
strain, and stress fields.