:atom:/:wrench: introduce model elements
Introduce model elements in their full form, including __str__ methods,
and add properties to each element that will provide the relevant input
data in the form used by the internal C routines. This form closely
resembles that originally used in Coulomb, for consistency at this
point. In the future, the C might be overhauled to compute these
elements with only specific data necessary for each type of element.
`raw_elements` was removed as a class attribute for the base `Model`
class.

Also change the transformer to enforce x-y ordering is consistent. Made
a number of adjustments throughout the package to ensure everything
matches.