:electron: add Askja deformation example
Add an example demonstrating how to use OkadaPy to model deformation
around Askja volcano, Iceland. This example, in Jupyter notebook form,
demonstrates how to build a model from scratch using the OkadaPy
interface, as well as how to use a pre-existing Coulomb input file
directly.

Note: I think a future goal should be to separate the 'model' from the
'x/y coords'—that is, the model should be a standalone entity that is
located somewhere in geographic space, but we should be able to then
'evaluate' that model at a set of x/y coordinate locations.